This is my first commit on this project. I hope to make this project span out years into the future. I wish to be able to make an impact on the wellbeing of other people as well as myself. I want to do good. I hope for a day when everyone can be happy and healthy. I dream of becoming a great nurse, salesman, informatician, and leader in my lifetime. I want to fall in love one day. I want to smile in the mirror again. I want to make others smile. I hope that the love I show others will one day be passed on, and I hope there will be a string of kindness that crosses generations because people chose to share that kindness I showed them, and then the next person, and then the next.
Hi James,

James, I hope you grow up to be man/woman/whatever who lives a great life. Follow your heart, lead with your mind, and keep your body in routine. Always. And if you stop, that's okay, but get back to it.
James, you got this. Aim to enjoy or love the process. Whatever that process may be.
James, I want you to know, you will grow up and come of age in a world of falsehoods and deceptions. Seek the truth. Seek it with an open mind, even to your enemies. For the truth will be the first step to setting yourself free.
James, Help yourself first before you help anyone else. A dead man can't hand out rations to refugees. And a wise man once said, 'first you get the money, then you get the power, then you get the [pretty monogomous lady who is totally down to commit or something like that]".

It's time for me to go to bed.

Goodnight James,
- Alex